bug-hunt #13: static analysis baseline - fix -Wconversion/-Wshadow/-Wpedantic across kernel
Fixes 20+ files with strict build warnings:
- idr.c: uint64_t→int conversion in idr_find()
- module_compress.c: uint64_t→uint32_t, int→uint16_t in LZMA decoder
- module_deps.c, mnt_namespace.c, igb.c: shadowed variable declarations
- pci.c, rtc.c, floppy.c, ps2.c: bitwise-and with narrowing casts
- e1000.c: label-before-declaration, function pointer cast
- vmxnet3.c: bit-field gen assignment narrowing
- rtl8139.c: int register offset narrowing
- acpi.c: sizeof arithmetic narrowing to uint32_t
- usb_audio.c: int→uint8_t for channel count
- partitions.c: uint64_t→uint32_t abs_start
- nbd.c, iscsi.c, nvmf.c, drbd.c, rbd.c: uint32_t→uint16_t TCP len
- fcoe.c: uint32_t→uint16_t in net_link_send
- cmos.c: BCD conversion int→uint8_t
- process.c: uint64_t→int in cap_bset_has()
- signal.c: void*→function pointer cast via union
- fat32.c: shadow bps, char overflow, format %zu→%llu
- fstab.c: shadow mounts
- fsck.c: uint64_t→uint32_t in blockdev calls
- iso9660.c, iso9660_rr.c: uint32_t→uint16_t uid/gid
- xattr.c: format %zu→%llu for custom size_t
- module.h: module_exit non-MODULE macro expansion

Create check-full baseline documenting -Wpadded warnings

+++ src/drivers/acpi.c
@@ -1685,13 +1685,13 @@ static int acpi_load_ssdts(void)
             kprintf("[ACPI] SSDT: expected XSDT signature\n");
             return 0;
         }
-        entry_count = (sdt_header->length - sizeof(struct acpi_header)) / 8;
+        entry_count = (uint32_t)((sdt_header->length - sizeof(struct acpi_header)) / 8);
         entry_size  = 8;
     } else if (g_rsdp_cache.rsdt_addr != 0) {
         sdt_header = (struct acpi_header *)PHYS_TO_VIRT((uint64_t)g_rsdp_cache.rsdt_addr);
         if (!sdt_header) return 0;
         if (memcmp(sdt_header->signature, "RSDT", 4) != 0) return 0;
-        entry_count = (sdt_header->length - sizeof(struct acpi_header)) / 4;
+        entry_count = (uint32_t)((sdt_header->length - sizeof(struct acpi_header)) / 4);
         entry_size  = 4;
     }
 
@@ -1840,7 +1840,7 @@ static void *acpi_find_table(const char signature[4])
             kprintf("[ACPI] find_table: XSDT checksum failed\n");
             return NULL;
         }
-        entry_count = (sdt_header->length - sizeof(struct acpi_header)) / 8;
+        entry_count = (uint32_t)((sdt_header->length - sizeof(struct acpi_header)) / 8);
         entry_size  = 8;
     } else if (g_rsdp_cache.rsdt_addr != 0) {
         /* Fall back to RSDT (32-bit physical addresses) */
@@ -1860,7 +1860,7 @@ static void *acpi_find_table(const char signature[4])
             kprintf("[ACPI] find_table: RSDT checksum failed\n");
             return NULL;
         }
-        entry_count = (sdt_header->length - sizeof(struct acpi_header)) / 4;
+        entry_count = (uint32_t)((sdt_header->length - sizeof(struct acpi_header)) / 4);
         entry_size  = 4;
     }
